Refactor code base.
