+++ test/lib/next-test-utils.js
@@ -5,6 +5,7 @@ import express from 'express'
 import path from 'path'
 import portfinder from 'portfinder'
 import { spawn } from 'child_process'
+import { readFileSync, writeFileSync, existsSync, unlinkSync } from 'fs'
 import fkill from 'fkill'
 
 import server from '../../dist/server/next'
@@ -154,5 +155,32 @@ export async function check (contentFn, regex) {
       if (regex.test(newContent)) break
       await waitFor(1000)
     } catch (ex) {}
+  }
+}
+
+export class File {
+  constructor (path) {
+    this.path = path
+    this.originalContent = existsSync(this.path) ? readFileSync(this.path, 'utf8') : null
+  }
+
+  write (content) {
+    if (!this.originalContent) {
+      this.originalContent = content
+    }
+    writeFileSync(this.path, content, 'utf8')
+  }
+
+  replace (pattern, newValue) {
+    const newContent = this.originalContent.replace(pattern, newValue)
+    this.write(newContent)
+  }
+
+  delete () {
+    unlinkSync(this.path)
+  }
+
+  restore () {
+    this.write(this.originalContent)
   }
 }
